// src/api/authentis-users/types.gen.ts
export type AddressDTO = {
    /**
     * city of address
     */
    city?: string;
};

export type CreateUser = {
    userName: string;
    email: string;
    password: string;
    firstName?: string;
    lastName?: string;
    birthDate?: string;
    address?: AddressDTO;
};